Move context.py from util to gloo
